updated to size_t for call arg count; array, map item counts

+++ src/gc/gcexpr.c
@@ -5,6 +5,7 @@
 #include "../include/token.h"
 #include "../include/utils.h"
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdio.h>
 
 // ===================
@@ -94,7 +95,7 @@ PExpr *NewLogical(Pgc *gc, PExpr *left, Token *op, PExpr *right) {
     return e;
 }
 
-PExpr *NewCallExpr(Pgc *gc, Token *op, PExpr *callee, PExpr **args, int count) {
+PExpr *NewCallExpr(Pgc *gc, Token *op, PExpr *callee, PExpr **args, size_t count) {
     PExpr *e = NewExpr(gc, EXPR_CALL, op);
     if (e == NULL) {
         return NULL;
@@ -106,7 +107,7 @@ PExpr *NewCallExpr(Pgc *gc, Token *op, PExpr *callee, PExpr **args, int count) {
     return e;
 }
 
-PExpr *NewArrayExpr(Pgc *gc, Token *op, PExpr **items, int count) {
+PExpr *NewArrayExpr(Pgc *gc, Token *op, PExpr **items, size_t count) {
     PExpr *e = NewExpr(gc, EXPR_ARRAY, op);
     if (e == NULL) {
         return NULL;
@@ -117,7 +118,7 @@ PExpr *NewArrayExpr(Pgc *gc, Token *op, PExpr **items, int count) {
     return e;
 }
 
-PExpr *NewMapExpr(Pgc *gc, Token *op, PExpr **items, int count) {
+PExpr *NewMapExpr(Pgc *gc, Token *op, PExpr **items, size_t count) {
     PExpr *e = NewExpr(gc, EXPR_MAP, op);
     if (e == NULL) {
         return NULL;
@@ -173,8 +174,8 @@ void FreeExpr(Pgc *gc, PExpr *e) {
     switch (e->type) {
         case EXPR_CALL: {
             FreeExpr(gc, e->exp.ECall.callee);
-            int count = e->exp.ECall.argCount;
-            for (int i = 0; i < count; i++) {
+            size_t count = e->exp.ECall.argCount;
+            for (size_t i = 0; i < count; i++) {
                 PExpr *ex = arrpop(e->exp.ECall.args);
                 FreeExpr(gc, ex);
             }
@@ -230,8 +231,8 @@ void FreeExpr(Pgc *gc, PExpr *e) {
             break;
         }
         case EXPR_ARRAY: {
-            int count = e->exp.EArray.count;
-            for (int i = 0; i < count; i++) {
+            size_t count = e->exp.EArray.count;
+            for (size_t i = 0; i < count; i++) {
                 FreeExpr(gc, arrpop(e->exp.EArray.items));
             }
             arrfree(e->exp.EArray.items);
@@ -241,7 +242,7 @@ void FreeExpr(Pgc *gc, PExpr *e) {
         }
         case EXPR_MAP: {
             struct EMap *map = &e->exp.EMap;
-            for (int i = 0; i < map->count; i++) {
+            for (size_t i = 0; i < map->count; i++) {
                 FreeExpr(gc, arrpop(map->etable));
             }
